// src/shared/api/schema.d.ts
export interface components {
  schemas: {
    /**
     * AccessTokenResponse
     * @description Represents a response containing access token and its expiration datetime.
     */
    AccessTokenResponse: {
      /** Access Token */
      access_token: string;
      /**
       * Token Type
       * @default bearer
       */
      token_type: string;
      /**
       * Expires At
       * Format: date-time
       */
      expires_at: string;
    };
    /** Body_login_api_v1_auth_login_post */
    Body_login_api_v1_auth_login_post: {
      /** Grant Type */
      grant_type?: string | null;
      /** Username */
      username: string;
      /** Password */
      password: string;
      /**
       * Scope
       * @default
       */
      scope: string;
      /** Client Id */
      client_id?: string | null;
      /** Client Secret */
      client_secret?: string | null;
    };
    /** Body_update_current_user_avatar_api_v1_users_me_avatar_patch */
    Body_update_current_user_avatar_api_v1_users_me_avatar_patch: {
      /**
       * File
       * Format: binary
       */
      file: Blob;
    };
    /**
     * CurrentUserEmailUpdateRequest
     * @description Represents a request to update current user's email.
     */
    CurrentUserEmailUpdateRequest: {
      /** Otp */
      otp: number;
      /**
       * Email
       * Format: email
       */
      email: string;
    };
    /**
     * CurrentUserResponse
     * @description Represents the private response data for a user.
     */
    CurrentUserResponse: {
      /**
       * Created At
       * Format: date-time
       */
      created_at: string;
      /**
       * Updated At
       * Format: date-time
       */
      updated_at: string;
      /** Id */
      id: number;
      /** Email */
      email: string;
      /** Avatar Url */
      avatar_url: string | null;
      /** Is Active */
      is_active: boolean;
      /** Is Verified */
      is_verified: boolean;
    };
    /** HTTPValidationError */
    HTTPValidationError: {
      /** Detail */
      detail?: components["schemas"]["ValidationError"][];
    };
    /**
     * OTPResponse
     * @description Represents a response containing One-Time Password (OTP) expiration datetime.
     */
    OTPResponse: {
      /**
       * Expires At
       * Format: date-time
       */
      expires_at: string;
    };
    /**
     * OTPVerifyRequest
     * @description Represents a request for verifying a One-Time Password (OTP).
     */
    OTPVerifyRequest: {
      /** Otp */
      otp: number;
      /**
       * Email
       * Format: email
       */
      email: string;
    };
    /**
     * UserEmailRequest
     * @description Represents a request containing a user's e-mail.
     */
    UserEmailRequest: {
      /**
       * Email
       * Format: email
       */
      email: string;
    };
    /**
     * UserExistenceResponse
     * @description Represents a response indicating whether a user exists.
     */
    UserExistenceResponse: {
      /** Exists */
      exists: boolean;
    };
    /**
     * UserPasswordRequest
     * @description Represents a request containing a user's password.
     */
    UserPasswordRequest: {
      /** Password */
      password: string;
    };
    /**
     * UserPasswordResetRequest
     * @description Represents a request for resetting a user's password.
     */
    UserPasswordResetRequest: {
      /** Otp */
      otp: number;
      /** Password */
      password: string;
      /**
       * Email
       * Format: email
       */
      email: string;
    };
    /**
     * UserRegistrationRequest
     * @description Represents a request for user registration.
     */
    UserRegistrationRequest: {
      /** Otp */
      otp: number;
      /** Password */
      password: string;
      /**
       * Email
       * Format: email
       */
      email: string;
    };
    /** ValidationError */
    ValidationError: {
      /** Location */
      loc: (string | number)[];
      /** Message */
      msg: string;
      /** Error Type */
      type: string;
    };
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}